Refactor functional tests for code sharing
The purpose of this refactoring is to clean up the code base a bit to
be able to share functional tests between DMAP and MRP. It is now
possible to have functional tests specific for each protocol but also
to share tests by adding them to common_functional_tests.py.
