+++ tests/airplay/test_airplay.py
@@ -1,13 +1,13 @@
-"""Functional tests using the API with a fake Apple TV."""
+"""Functional tests for Airplay."""
 
 import asyncio
 
-from tests.log_output_handler import LogOutputHandler
 from aiohttp import ClientSession
 from aiohttp.test_utils import (AioHTTPTestCase, unittest_run_loop)
 
 from pyatv.airplay import player
-from tests.fake_daap_atv import (FakeDaapAppleTV, AppleTVUseCases)
+from tests.airplay.fake_airplay_device import (
+    FakeAirPlayDevice, AirPlayUseCases)
 
 
 STREAM = 'http://airplaystream'
@@ -18,7 +18,6 @@ class AirPlayPlayerTest(AioHTTPTestCase):
 
     def setUp(self):
         AioHTTPTestCase.setUp(self)
-        self.log_handler = LogOutputHandler(self)
 
         # This is a hack that overrides asyncio.sleep to avoid making the test
         # slow. It also counts number of calls, since this is quite important
@@ -26,15 +25,11 @@ class AirPlayPlayerTest(AioHTTPTestCase):
         player.asyncio.sleep = self.fake_asyncio_sleep
         self.no_of_sleeps = 0
 
-    def tearDown(self):
-        AioHTTPTestCase.tearDown(self)
-        self.log_handler.tearDown()
-
     @asyncio.coroutine
     def get_application(self, loop=None):
-        self.fake_atv = FakeDaapAppleTV(self.loop, 0, 0, 0, self)
-        self.usecase = AppleTVUseCases(self.fake_atv)
-        return self.fake_atv
+        self.fake_device = FakeAirPlayDevice(self.loop, self)
+        self.usecase = AirPlayUseCases(self.fake_device)
+        return self.fake_device
 
     @asyncio.coroutine
     def fake_asyncio_sleep(self, time, loop):
@@ -51,8 +46,8 @@ class AirPlayPlayerTest(AioHTTPTestCase):
             self.loop, session, '127.0.0.1', port=self.server.port)
         yield from aplay.play_url(STREAM, position=START_POSITION)
 
-        self.assertEqual(self.fake_atv.last_airplay_url, STREAM)
-        self.assertEqual(self.fake_atv.last_airplay_start, START_POSITION)
+        self.assertEqual(self.fake_device.last_airplay_url, STREAM)
+        self.assertEqual(self.fake_device.last_airplay_start, START_POSITION)
         self.assertEqual(self.no_of_sleeps, 2)  # playback + idle = 3
 
         session.close()
