+++ tox.ini
@@ -17,7 +17,7 @@ ignore_errors = True
 commands =
      flake8 --exclude=pyatv/mrp/protobuf pyatv tests examples
      pylint pyatv examples
-     pydocstyle -v --match='(?!test_).*[^pb2]\.py' pyatv examples/ tests
+     pydocstyle -v --match='(?!test_).*[^pb2]\.py' pyatv examples
 
 [testenv:typing]
 basepython = python3
